backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
tensorflow==2.18.1
pillow==9.5.0
numpy
gunicorn==20.1.0
python-dotenv==1.0.0
